Moving old bookmark manager to LegacyBookmarkManager
LegacyBookmarkManager has read only functionality
